requirements.txt
asgiref==3.8.1
asttokens==2.4.1
colorama==0.4.6
crispy-bootstrap5==2024.2
decorator==5.1.1
Django==5.0.7
django-autoslug==1.9.9
django-crispy-forms==2.3
django-extensions==3.2.3
django-tinymce==4.1.0
easy-thumbnails==2.9
executing==2.0.1
ipython==8.26.0
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pillow==10.4.0
pip==24.1.2
prompt_toolkit==3.0.47
pure_eval==0.2.3
Pygments==2.18.0
python-slugify==8.0.4
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
text-unidecode==1.3
traitlets==5.14.3
tzdata==2024.1
wcwidth==0.2.13
